requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
Django
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
pillow==12.1.0
psycopg==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
SQLAlchemy==2.0.46
sqlparse==0.5.5
typing_extensions==4.15.0
Werkzeug==3.1.5